fixes issue #4490, fixes issue #6237 - changes the dlineedit to allow entering and displaying dates in the user's current locale and pops up a calendar widget for selecting dates. also fixes some postgresql 8.3 problems encountered during testing, incorporates some code modernizations, and moves storedProcErrorLookup from guiclient to common so it can be used in the batchItem window shared by the guiclient and other programs
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@1988 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/invoice.cpp
@@ -1121,7 +1121,7 @@ void invoice::populateCMInfo()
 	      " WHERE ( (aropenco_cohead_id=cohead_id)"
 	      "   AND   (aropenco_aropen_id=aropen_id)"
 	      "   AND   (cohead_number=:cohead_number) );");
-    cm.bindValue(":cohead_number", _orderNumber->text().toInt());
+    cm.bindValue(":cohead_number", _orderNumber->text());
   }
   else
   {
@@ -1190,7 +1190,7 @@ void invoice::populateCCInfo()
 	      "   AND   (payco_ccpay_id=ccpay_id)"
 	      "   AND   (payco_cohead_id=cohead_id)"
 	      "   AND   (cohead_number=:cohead_number) );");
-    cc.bindValue(":cohead_number", _orderNumber->text().toInt());
+    cc.bindValue(":cohead_number", _orderNumber->text());
   }
   else
   {
